Move Registry to the model bundle API
