+++ aws/aws-mcp-cli-commands/src/main/resources/META-INF/services/software.amazon.smithy.mcp.bundle.api.Registry
